test: move tests out of reth-primitives (#13636)
